Feat: go 여부를  인터페이스로 구현
