+++ src/main/java/car/step3/HandleCar.java
@@ -3,6 +3,7 @@ package car.step3;
 import java.util.List;
 
 public class HandleCar {
+    MovingStrategy movingStrategy = new RandomMovingStrategy();
 
     public List<Car> iteratorCar(List<Car> cars) {
         for (Car car : cars) {
@@ -12,6 +13,6 @@ public class HandleCar {
     }
 
     public void goCar(Car car) {
-        car.go();
+        car.go(movingStrategy);
     }
 }
