+++ src/main/java/car/step3/MovingStrategy.java
@@ -1,0 +1,5 @@
+package car.step3;
+
+public interface MovingStrategy {
+    boolean movable();
+}
