+++ src/main/java/car/step3/RandomMovingStrategy.java
@@ -1,0 +1,17 @@
+package car.step3;
+
+import java.util.Random;
+
+public class RandomMovingStrategy implements MovingStrategy{
+    private int MAX_BOUNCE = 10;
+    private int FORWARD_NUMBER = 4;
+    @Override
+    public boolean movable() {
+        return getRandomNumber() >= FORWARD_NUMBER;
+    }
+
+    protected int getRandomNumber() {
+        Random random = new Random();
+        return random.nextInt(MAX_BOUNCE);
+    }
+}
