+++ src/test/java/car/step3/CarTest.java
@@ -12,27 +12,22 @@ class CarTest {
     HandleCar handleCar = new HandleCar();
     @Test
     public void 차가_전진하는지_테스트() {
-        Car car = new Car("new") {
-            @Override
-            protected int getRandomNumber() {
-                return 4;
-            }
-        };
+        Car car = new Car("lim");
 
-        car.go();
+        car.go(new MovingStrategy() {
+            @Override
+            public boolean movable() {
+                return true;
+            }
+        });
         assertThat(car.location()).isEqualTo(1);
     }
 
     @Test
     public void 차가_정지하는지_확인하는_테스트() {
-        Car car = new Car("not_moved") {
-            @Override
-            protected int getRandomNumber() {
-                return 3;
-            }
-        };
+        Car car = new Car("kim");
 
-        car.go();
+        car.go(() -> false);
         assertThat(car.location()).isEqualTo(0);
     }
